Rename batches/batches.ts -> batches/index.ts
