+++ client/web/src/org/area/OrgHeader.tsx
@@ -4,7 +4,7 @@ import { Link, NavLink, RouteComponentProps } from 'react-router-dom'
 
 import { PageHeader } from '@sourcegraph/wildcard'
 
-import { BatchChangesProps } from '../../batches/batches'
+import { BatchChangesProps } from '../../batches'
 import { NavItemWithIconDescriptor } from '../../util/contributions'
 import { OrgAvatar } from '../OrgAvatar'
