+++ client/web/src/routes.tsx
@@ -3,7 +3,7 @@ import { Redirect, RouteComponentProps } from 'react-router'
 
 import { isErrorLike } from '@sourcegraph/shared/src/util/errors'
 
-import { BatchChangesProps } from './batches/batches'
+import { BatchChangesProps } from './batches'
 import { BreadcrumbsProps, BreadcrumbSetters } from './components/Breadcrumbs'
 import type { LayoutProps } from './Layout'
 import type { ExtensionAlertProps } from './repo/RepoContainer'
